+++ src/Lifecycle/AbstractProcessEvent.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace Paraunit\Lifecycle;
 
 use Paraunit\Process\AbstractParaunitProcess;
-use Symfony\Component\EventDispatcher\Event;
 
-abstract class AbstractProcessEvent extends Event
+abstract class AbstractProcessEvent extends AbstractEvent
 {
     /** @var AbstractParaunitProcess */
     private $process;