+++ src/Lifecycle/ForwardCompatEventDispatcher.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Paraunit\Lifecycle;
 
-use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -18,7 +17,7 @@ class ForwardCompatEventDispatcher implements EventDispatcherInterface
         $this->eventDispatcher = $eventDispatcher;
     }
 
-    public function dispatch($event, Event $dummyEventArgument = null)
+    public function dispatch($event, AbstractEvent $dummyEventArgument = null)
     {
         if (! \is_object($event)) {
             throw new \InvalidArgumentException('Expecting object event, got ' . \gettype($event));